pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openai_tt"
version = "0.1.0"
description = "Real-time audio transcription and translation using OpenAI API"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "python-dotenv",
    "websockets",
    "sounddevice",
    "asyncstdlib",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[tool.setuptools]
py-modules = ["main", "audio"]